requirements.txt
chemotion-converter-app @ git+https://github.com/ComPlat/chemotion-converter-app
mdutils ~= 1.7.0
click ~= 8.2.1